edit: 에러핸들링

+++ src/posts/post.controller.js
@@ -1,4 +1,5 @@
 const PostService = require('./post.service')
+const CustomError = require('../middlewares/error.middleware')
 
 class PostController {
   postService = new PostService()
@@ -27,7 +28,8 @@ class PostController {
     //       storageClass: 'STANDARD',
     //       serverSideEncryption: null,
     //       metadata: null,
-    //       location: 'https://your-bucket-name.s3.amazonaws.com/uploads/image1_1.jpg', // S3 주소
+    //       location: 'https://your-bucket-name.s3.amazonaws.com/images/${Date.now().toString}_${file.originalname}.jpg', // S3 주소
+    //                 image.path === ${Date.now().toString}_${file.originalname}.jpg
     //       etag: '"abcdefgh1234567890"',
     //       versionId: undefined
     //     }
@@ -37,6 +39,11 @@ class PostController {
     // }
 
     try {
+      // 로그인 유무 확인
+      if (!userId) {
+        throw new CustomError('로그인 후 사용가능합니다.', 401)
+      }
+
       // 게시물 생성
       const createPostWithImage = await this.postService.createPostWithImage(
         userId,
@@ -60,6 +67,8 @@ class PostController {
   findPostWithImage = async (req, res, next) => {
     const { postId } = req.params
     try {
+      // 게시글 삭제 권한 확인
+
       // auth.middleware에서 값을 받아야 함
       let { userId } = res.locals.user
       
@@ -84,6 +93,15 @@ class PostController {
     const images = req.files
     
     try {
+      // 게시글 수정 권한 확인
+      const confirmRight = this.postService.confirmRight(postId, userId)
+
+      // 권한이 없다면 errorMessage
+      if (!confirmRight) {
+        throw new CustomError('수정 권한이 없습니다.', 403)
+      }
+
+      // 게시글 수정
       const updatePostWithImage = this.postService.updatePostWithImage( // {updatePost, updateImage}
         postId,
         userId,
@@ -94,8 +112,9 @@ class PostController {
         images
       )
 
+      // 수정 실패 시 errorMessage
       if(!updatePostWithImage.updatePost.length) {
-        res.status(500).json({ errorMessage: '수정에 실패하였습니다.'})
+        throw new CustomError('수정에 실패했습니다.', 500)
       }
 
       res.status(200).json({ message: '게시글이 수정되었습니다.'})
@@ -111,9 +130,14 @@ class PostController {
     const { postId } = req.params
     const { userId } = res.locals.user
     
-
     try {
-      // 게시글 삭제 권한 확인 ( 추가 예정 )
+      // 게시글 삭제 권한 확인
+      const confrimRight = await this.postService.confrimRight(postId, userId)
+
+      // 권한이 없다면 errorMessage
+      if (!confrimRight) {
+        throw new CustomError('삭제 권한이 없습니다.', 403)
+      }
 
       // 게시글 삭제
       const deletePostWithImage = await this.postService.deletePostWithImage(
@@ -123,8 +147,7 @@ class PostController {
       
       // 게시글 삭제 실패 시 error
       if (!deletePostWithImage) {
-        res.status(500).json({ errorMessage: '게시물 삭제에 실패했습니다.'})
-        return
+        throw new CustomError('게시물 삭제 실패했습니다.', 500)
       }
 
       res.status(200).json({ message: '게시물 삭제에 성공했습니다.'})
